requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.1
networkx==3.4.2
numpy==1.26.4
openai==1.60.2
pandas==2.2.3
pydantic==2.11.7
pydantic_core==2.33.2
pysam==0.22.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.3.2
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
